+++ src/AppBundle/Controller/ProductController.php
@@ -8,6 +8,7 @@ use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -17,44 +18,54 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
  * Class ProductController
  * @package AppBundle\Controller
  * @Route("/products")
- * @Security("has_role('ROLE_MODERATOR')")
  */
 class ProductController extends Controller
 {
     /**
+     * @param Request $request
      * @return Response
      * @Route("/", name="products")
+     * @Security("has_role('ROLE_USER')")
      */
-    public function indexAction()
+    public function indexAction(Request $request)
     {
-        $em = $this->getDoctrine()->getManager();
-        $products = $em->getRepository('AppBundle:Product')->findAll();
+        if ($request->isXmlHttpRequest()) {
+            $response = $this->get('filter_service')->getByFilters($request, 'AppBundle:Product', [
+                'createDate',
+                'updateDate',
+                'sku'
+            ]);
+
+            return $response;
+        }
 
         return $this->render('products/index.html.twig', array(
-            'products' => $products,
+            'filter' => '',
         ));
     }
 
     /**
      * @param Request $request
-     * @return RedirectResponse|Response
+     * @return RedirectResponse|Response|JsonResponse
      * @Route("/create", name="product_create")
+     * @Security("has_role('ROLE_MODERATOR')")
      */
     public function createAction(Request $request)
     {
-        $product = new Product();
-        $form = $this->createForm(EditProductType::class, $product);
+        $form = $this->createForm(EditProductType::class);
         $form->handleRequest($request);
 
+        if ($request->isXmlHttpRequest()) {
+            $name = $request->request->get('name');
+            $sku = $request->request->get('sku');
+
+            return $this->get('products')->isExistProductAdd($name, $sku);
+        }
+
         if ($form->isSubmitted() && $form->isValid()) {
-            $result = $this->get('products')->createProduct($form);
+            $this->get('products')->createProduct($form);
 
-            if (!$result) {
-                $this->addFlash('error', 'Product is already created');
-                return $this->redirectToRoute('product_create');
-            }
-
-            $this->addFlash('message', 'Product was created');
+            $this->addFlash('message', 'Продукт был успешно создан.');
             return $this->redirectToRoute('products');
         }
 
@@ -70,6 +81,7 @@ class ProductController extends Controller
      *     requirements={"id" = "\d+"},
      *     defaults={"id" = 1},
      *     name="product_details")
+     * @Security("has_role('ROLE_USER')")
      */
     public function detailsAction($id)
     {
@@ -77,7 +89,7 @@ class ProductController extends Controller
         $product = $em->getRepository('AppBundle:Product')->find($id);
 
         if (is_null($product)) {
-            throw new NotFoundHttpException();
+            throw new NotFoundHttpException('Продукт не найден');
         }
 
         return $this->render('products/details.html.twig', array(
@@ -93,6 +105,7 @@ class ProductController extends Controller
      *     requirements={"id" = "\d+"},
      *     defaults={"id" = 1},
      *     name="product_edit_id")
+     * @Security("has_role('ROLE_MODERATOR')")
      */
     public function editAction(Request $request, $id)
     {
@@ -100,50 +113,29 @@ class ProductController extends Controller
         $product = $em->getRepository('AppBundle:Product')->find($id);
 
         if (is_null($product)) {
-            throw new NotFoundHttpException('Product not found');
+            throw new NotFoundHttpException('Продукт не найден');
         }
 
-        $form = $this->createForm(EditProductType::class, $product);
+        $form = $this->createForm(EditProductType::class);
+        $this->get('products')->fillFormWithDataOfProduct($form, $product);
         $form->handleRequest($request);
 
+        if ($request->isXmlHttpRequest()) {
+            $name = $request->request->get('name');
+            $sku = $request->request->get('sku');
+
+            return $this->get('products')->isExistProductEdit($name, $sku, $product);
+        }
+
         if ($form->isSubmitted() && $form->isValid()) {
-            $result = $this->get('products')->editProduct($form);
+            $this->get('products')->editProduct($form, $product);
 
-            if (!$result) {
-                $this->addFlash('error', 'Product is already created');
-                return $this->redirectToRoute('product_edit_id', array(
-                    'id' => $id,
-                ));
-            }
-
-            $this->addFlash('message', 'Product was changed');
+            $this->addFlash('message', 'Продукт был успешно изменен');
             return $this->redirectToRoute('products');
         }
 
         return $this->render('products/edit.html.twig', array(
             'form' => $form->createView(),
-        ));
-    }
-
-    /**
-     * @param $id
-     * @return Response
-     * @Route("/delete/{id}",
-     *     requirements={"id" = "\d+"},
-     *     defaults={"id" = 1},
-     *     name="product_delete")
-     * @Method({"GET"})
-     */
-    public function deleteAction($id)
-    {
-        $em = $this->getDoctrine()->getManager();
-        $product = $em->getRepository('AppBundle:Product')->find($id);
-
-        if (is_null($product)) {
-            throw new NotFoundHttpException('Product not found');
-        }
-
-        return $this->render('products/delete.html.twig', array(
             'product' => $product,
         ));
     }
@@ -155,16 +147,97 @@ class ProductController extends Controller
      *     requirements={"id" = "\d+"},
      *     defaults={"id" = 1})
      * @Method({"POST"})
+     * @Security("has_role('ROLE_MODERATOR')")
      */
     public function deleteConfirmAction($id)
     {
         $em = $this->getDoctrine()->getManager();
         $product = $em->getRepository('AppBundle:Product')->find($id);
 
+        if (is_null($product)) {
+            throw new NotFoundHttpException('Продукт не найден');
+        }
+
         $em->remove($product);
         $em->flush();
 
-        $this->addFlash('message', 'Product was deleted');
+        $this->addFlash('message', 'Продукт был успешно удален');
         return $this->redirectToRoute('products');
+    }
+
+    /**
+     * @param Request $request
+     * @return Response
+     * @Route("/get_all_categories", name="get_all_categories")
+     * @Method({"POST"})
+     */
+    public function getAllCategoriesAction(Request $request)
+    {
+        if ($request->isXmlHttpRequest()) {
+            $em = $this->getDoctrine()->getManager();
+            $categories = $em->getRepository('AppBundle:Category')->findAll();
+
+            $response = $this->get('serialize_service')->serializeObjects($categories);
+
+            return $response;
+        }
+    }
+
+    /**
+     * @param Request $request
+     * @return Response
+     * @Route("/get_all_manufacturers", name="get_all_manufacturers")
+     * @Method({"POST"})
+     */
+    public function getAllManufacturersAction(Request $request)
+    {
+        if ($request->isXmlHttpRequest()) {
+            $em = $this->getDoctrine()->getManager();
+            $manufacturers = $em->getRepository('AppBundle:Manufacturer')->findAll();
+
+            $response = $this->get('serialize_service')->serializeObjects($manufacturers);
+
+            return $response;
+        }
+    }
+
+    /**
+     * @param Request $request
+     * @return Response
+     * @Route("/get_count")
+     */
+    public function getCountRows(Request $request)
+    {
+        if ($request->isXmlHttpRequest()) {
+            $length = $this->get('filter_service')->getCountRows($request, 'AppBundle:Product');
+
+            return $length;
+        }
+    }
+
+    /**
+     * @param Request $request
+     * @return Response
+     * @Route("/get_by_filter", name="get_by_filter")
+     */
+    public function getByFilter(Request $request)
+    {
+        if ($request->isXmlHttpRequest()) {
+            $filter = $request->get('filter');
+
+            $em = $this->getDoctrine()->getManager();
+            $products = $em->getRepository('AppBundle:Product')
+                ->findByFilter($filter);
+
+            $response = $this->get('serialize_service')->serializeObjects($products);
+
+            return $response;
+        }
+
+        $filter = $request->get('searchInput');
+
+        return $this->render('products/index.html.twig', array(
+            'filter' => $filter,
+        ));
     }
 }
